+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 * Minify CSS file
 * Loading icon with Font Awesome instead of Glyphicon
+* New setting: Dialog Insight Bookings Number Field
 
 ### 2.0.4: April 6th, 2022
 * Send to Aquos only bookings "self", exclude for "other"